Improve archipelago UI usability
