+++ Randomiser/Multiplayer/Archipelago/ArchipelagoItem.cs
@@ -1,0 +1,14 @@
+﻿namespace Randomiser.Multiplayer.Archipelago
+{
+    public struct ArchipelagoItem
+    {
+        public readonly string name;
+        public readonly string key;
+
+        public ArchipelagoItem(string name, string key)
+        {
+            this.name = name;
+            this.key = key;
+        }
+    }
+}
